CC-158 PoC move color control from Annotation to Rendering for annotation layers
